+++ cybroscgiserver/rootfs/etc/cont-init.d/cybroscgiserver.sh
@@ -5,6 +5,11 @@
 # ==============================================================================
 # Creates initial CybroScgiServer configuration in case it is non-existing
 declare configuration_file
+declare autodetect_address
+declare request_period_s
+declare valid_period_s
+declare cleanup_period_s
+declare verbose_level
 
 configuration_file=$(bashio::config 'configuration_file')
 if ! bashio::fs.file_exists "${configuration_file}"; then
@@ -20,7 +25,35 @@ if ! bashio::fs.file_exists "${configuration_file}"; then
     bashio::log.fatal
     bashio::exit.nok
 else
-    bashio::log.info "Using existing configuration file:"
-    bashio::log.info "${configuration_file}"
+    bashio::log.info "Using existing configuration file: ${configuration_file}"
+    cp "$(bashio::config 'configuration_file')" /usr/local/bin/scgi_server/config.ini
+
+    if bashio::config.has_value "autodetect_address"; then autodetect_address=$(bashio::config 'autodetect_address'); else autodetect_address=""; fi
+    if [ "$autodetect_address" != "" ]; then
+        bashio::log.info "configured autodetect address: ${autodetect_address}"
+        sed -i "s/^autodetect_enabled.*/autodetect_enabled = true/" /usr/local/bin/scgi_server/config.ini
+        sed -i "s/^autodetect_address.*/autodetect_address = $autodetect_address/" /usr/local/bin/scgi_server/config.ini
+    else
+        bashio::log.warning "Found no autodetect address in config.ini!"
+        bashio::log.warning "Controller autodetect is not available!"
+        bashio::log.warning "To use autodetect you need to setup controllers"
+        bashio::log.warning "manually in config.ini file!"
+        sed -i "s/^autodetect_enabled.*/autodetect_enabled = false/" /usr/local/bin/scgi_server/config.ini
+    fi
+
+    if bashio::config.has_value "request_period_s"; then request_period_s=$(bashio::config 'request_period_s'); else request_period_s="0"; fi
+    bashio::log.info "configured request_period_s: ${request_period_s}"
+    sed -i "s/^request_period_s.*/request_period_s = $request_period_s/" /usr/local/bin/scgi_server/config.ini
+    if bashio::config.has_value "valid_period_s"; then valid_period_s=$(bashio::config 'valid_period_s'); else valid_period_s="0"; fi
+    bashio::log.info "configured valid_period_s: ${valid_period_s}"
+    sed -i "s/^valid_period_s.*/valid_period_s = $valid_period_s/" /usr/local/bin/scgi_server/config.ini
+    if bashio::config.has_value "cleanup_period_s"; then cleanup_period_s=$(bashio::config 'cleanup_period_s'); else cleanup_period_s="0"; fi
+    bashio::log.info "configured cleanup_period_s: ${cleanup_period_s}"
+    sed -i "s/^cleanup_period_s.*/cleanup_period_s = $cleanup_period_s/" /usr/local/bin/scgi_server/config.ini
+    if bashio::config.has_value "verbose_level"; then verbose_level=$(bashio::config 'verbose_level'); else verbose_level="ERROR"; fi
+    bashio::log.info "configured verbose_level: ${verbose_level}"
+    sed -i "s/^verbose_level.*/verbose_level = $verbose_level/" /usr/local/bin/scgi_server/config.ini
+
+    cp /usr/local/bin/scgi_server/config.ini "$(bashio::config 'configuration_file')"
     bashio::exit.ok
 fi